Volume: don't use a block allocator when mounting readonly. BitmapBlock: also use file system blocks, current type is off_t. Also added more trace. BlockAllocator: added an assert and more trace
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@39302 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/add-ons/kernel/file_systems/ext2/BitmapBlock.h
@@ -16,9 +16,9 @@ public:
 							BitmapBlock(Volume* volume, uint32 numBits);
 							~BitmapBlock();
 
-			bool			SetTo(uint32 block);
+			bool			SetTo(off_t block);
 			bool			SetToWritable(Transaction& transaction,
-								uint32 block, bool empty = false);
+								off_t block, bool empty = false);
 
 			bool			CheckMarked(uint32 start, uint32 length);
 			bool			CheckUnmarked(uint32 start, uint32 length);
